+++ sdk/builder.go
@@ -1,9 +1,11 @@
 package sdk
 
+import "github.com/XingfenD/rainyun_api_go_sdk/apis"
+
 // SDKBuilder constructs a RainyunSDK fluently.
 type SDKBuilder struct {
 	apiKey string
-	trace  *TraceOptions
+	trace  *apis.TraceOptions
 }
 
 func NewBuilder(apiKey string) *SDKBuilder {
@@ -12,14 +14,11 @@ func NewBuilder(apiKey string) *SDKBuilder {
 
 // WithTrace enables structured tracing with the given options. Passing nil
 // disables tracing.
-func (b *SDKBuilder) WithTrace(options *TraceOptions) *SDKBuilder {
+func (b *SDKBuilder) WithTrace(options *apis.TraceOptions) *SDKBuilder {
 	b.trace = options
 	return b
 }
 
 func (b *SDKBuilder) Build() *RainyunSDK {
-	if b.trace == nil {
-		return New(b.apiKey)
-	}
-	return NewWithTrace(b.apiKey, *b.trace)
+	return newSDK(apis.NewBuilder(b.apiKey).WithTrace(b.trace).Build())
 }